Added WEBAPP_PARAM fields, updated webapp_str_t definition to use long long.
Removed DisableBackgroundQueue. Use SetParamInt(WEBAPP_PARAM_BGQUEUE, 0) instead.
Made GetTemplate not require page name.
TODO: Implement filesystem iterator/GetFiles API.
TODO: Implement refresh_templates in init.lua.

+++ plugins/core/handlers.lua
@@ -1,6 +1,6 @@
 local ffi = require("ffi")
 ffi.cdef[[
-typedef struct { const char* data; int len; } webapp_str_t;
+typedef struct { const char* data; long long len; } webapp_str_t;
 int QueueProcess(void* app, webapp_str_t* func, webapp_str_t* vars);
 void ClearCache(void* app, void* requests);
 void DestroySession(void* session);
